Seperated the task 2.1 functions to their own file
